Move css files to 'style' folder
